refactor: move the Products components for store panel page to another folder to match the project structure
