+++ src/components/storePanel/products/ProductsList/ProductItemDesktop.tsx
