+++ src/pages/storePanel/Products/index.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
 import { useWindowDimensions } from '@/hooks/useWindowDimensions';
 import { productStatusList } from '@/constants/productStatusList';
-import { ProductsList } from '@/components/storePanel/ProductsList';
+import { ProductsList } from '@/components/storePanel/products/ProductsList';
 import { AddProductButton } from '@/components/storePanel/AddProductButton';
 import { EmptyContentPage } from '@/components/EmptyContentPage';
-import { ProductSorting } from '@/components/storePanel/ProductSorting';
+import { ProductSorting } from '@/components/storePanel/products/ProductSorting';
 import { Icon } from '@iconify/react';
 import s from './Products.module.scss';
